Remove featuretogglesprovider and use sykefraværsstatistikk as hook
