+++ src/hooks/useFeatureToggles.ts
@@ -1,12 +1,8 @@
-import React, { createContext, FunctionComponent, useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { hentRestFeatureToggles, RestFeatureToggles } from '../api/feature-toggles-api';
 import { RestStatus } from '../api/api-utils';
 
-export const featureTogglesContext = createContext<RestFeatureToggles>({
-    status: RestStatus.LasterInn,
-});
-
-export const FeatureTogglesProvider: FunctionComponent = (props) => {
+export function useFeatureToggles() {
     const [featureToggles, setFeatureToggles] = useState<RestFeatureToggles>({
         status: RestStatus.LasterInn,
     });
@@ -16,7 +12,7 @@ export const FeatureTogglesProvider: FunctionComponent = (props) => {
             setFeatureToggles(
                 await hentRestFeatureToggles(
                     'sykefravarsstatistikk.ab-test.tips', // unleash: https://unleash.nais.io/#/features/strategies/sykefravarsstatistikk.ab-test.tips
-                    'sykefravarsstatistikk.arbeidsmiljoportal', // unleash: https://unleash.nais.io/#/features/strategies/sykefravarsstatistikk.arbeidsmiljoportal
+                    'sykefravarsstatistikk.arbeidsmiljoportal' // unleash: https://unleash.nais.io/#/features/strategies/sykefravarsstatistikk.arbeidsmiljoportal
                     /* Send med features her */
                 )
             );
@@ -24,6 +20,5 @@ export const FeatureTogglesProvider: FunctionComponent = (props) => {
 
         hentFeatureTogglesOgSettState();
     }, []);
-    const Provider = featureTogglesContext.Provider;
-    return <Provider value={featureToggles}>{props.children}</Provider>;
-};
+    return featureToggles;
+}
