+++ src/hooks/useSykefravarsstatistikk.ts
@@ -6,6 +6,10 @@ import { useVirksomhetsdata } from './useVirksomhetsdata';
 import { EnhetsregisteretState, useEnheter } from './useEnheter';
 import { RestSummertSykefraværshistorikk } from '../api/summert-sykefraværshistorikk-api';
 import { useSummertSykefravær } from './useSummertSykefravær';
+import { useSykefraværshistorikk } from './useSykefraværshistorikk';
+import { RestSykefraværshistorikk } from '../api/kvartalsvis-sykefraværshistorikk-api';
+import { useFeatureToggles } from './useFeatureToggles';
+import { RestFeatureToggles } from '../api/feature-toggles-api';
 
 export interface Sykefravarsstatistikk {
     altinnOrganisasjoner: RestAltinnOrganisasjoner;
@@ -13,6 +17,8 @@ export interface Sykefravarsstatistikk {
     virksomhetsdata: RestVirksomhetsdata;
     enhetsInfomasjon: EnhetsregisteretState;
     summertSykefravær: RestSummertSykefraværshistorikk;
+    fraværshistorikk: RestSykefraværshistorikk;
+    featureToggles: RestFeatureToggles;
 }
 
 export function useSykefravarsstatistikk(): Sykefravarsstatistikk {
@@ -21,11 +27,15 @@ export function useSykefravarsstatistikk(): Sykefravarsstatistikk {
     const virksomhetsdata = useVirksomhetsdata();
     const enhetsInfomasjon = useEnheter();
     const summertSykefravær = useSummertSykefravær();
+    const fraværshistorikk = useSykefraværshistorikk();
+    const featureToggles = useFeatureToggles();
     return {
         altinnOrganisasjoner,
         altinnOrganisasjonerMedStatistikk,
         virksomhetsdata,
         enhetsInfomasjon,
         summertSykefravær,
+        fraværshistorikk,
+        featureToggles,
     };
 }
